+++ packages/server/src/object.rs
@@ -1,3 +1,4 @@
+mod complete;
 mod get;
 mod metadata;
 mod put;
