+++ packages/server/src/object/complete.rs
@@ -1,0 +1,38 @@
+use crate::Server;
+use indoc::formatdoc;
+use tangram_client as tg;
+use tangram_database::{self as db, Database as _, Query as _};
+
+impl Server {
+	pub(crate) async fn try_get_object_complete_local(
+		&self,
+		id: &tg::object::Id,
+	) -> tg::Result<Option<bool>> {
+		// Get an index connection.
+		let connection = self
+			.index
+			.connection()
+			.await
+			.map_err(|source| tg::error!(!source, "failed to get a database connection"))?;
+
+		// Get the object metadata.
+		let p = connection.p();
+		let statement = formatdoc!(
+			"
+				select complete
+				from objects
+				where id = {p}1;
+			",
+		);
+		let params = db::params![id];
+		let output = connection
+			.query_optional_into(statement.into(), params)
+			.await
+			.map_err(|source| tg::error!(!source, "failed to execute the statement"))?;
+
+		// Drop the database connection.
+		drop(connection);
+
+		Ok(output)
+	}
+}
